Align online orders workspace
