+++ packages/athena-webapp/src/components/operations/DailyCloseView.tsx
@@ -32,7 +32,6 @@ import {
   type NormalizedApprovalCommandResult,
   type NormalizedCommandResult,
 } from "@/lib/errors/runCommand";
-import { formatStoredCurrencyAmount } from "@/lib/pos/displayAmounts";
 import { getOrigin } from "@/lib/navigationUtils";
 import { api } from "~/convex/_generated/api";
 import type { Id } from "~/convex/_generated/dataModel";
@@ -73,10 +72,13 @@ import {
 } from "./useApprovedCommand";
 import { OperationReviewWorkspace } from "./OperationReviewWorkspace";
 import { OperationReviewItemCard } from "./OperationReviewItemCard";
+import { OperationsSummaryMetric } from "./OperationsSummaryMetric";
+import { formatOperationsMetricHelper } from "./operationsMetricFormatting";
 import {
-  formatOperationsMetricHelper,
-  OperationsSummaryMetric,
-} from "./OperationsSummaryMetric";
+  formatDailyCloseCompletedAt,
+  formatDailyCloseMoney,
+  formatDailyCloseOperatingDate,
+} from "./dailyCloseFormatting";
 
 type DailyCloseApi = {
   completeDailyClose?: unknown;
@@ -614,17 +616,6 @@ function sentenceFragment(value: string) {
   return value ? value.charAt(0).toLocaleLowerCase() + value.slice(1) : value;
 }
 
-export function formatDailyCloseMoney(
-  currency: string,
-  amount?: number | null,
-) {
-  if (typeof amount !== "number") return "Pending";
-
-  return formatStoredCurrencyAmount(currency, amount, {
-    revealMinorUnits: true,
-  });
-}
-
 function DailyCloseFinancialValue({
   amount,
   canView,
@@ -641,21 +632,6 @@ function DailyCloseFinancialValue({
       {formatDailyCloseMoney(currency, amount)}
     </FinancialValue>
   );
-}
-
-export function formatDailyCloseOperatingDate(operatingDate: string) {
-  const parsed = new Date(`${operatingDate}T00:00:00`);
-
-  if (Number.isNaN(parsed.getTime())) {
-    return operatingDate;
-  }
-
-  return parsed.toLocaleDateString([], {
-    day: "numeric",
-    month: "short",
-    weekday: "long",
-    year: "numeric",
-  });
 }
 
 function getLocalDateFromOperatingDate(operatingDate: string) {
@@ -675,18 +651,6 @@ function getLocalDateFromOperatingDate(operatingDate: string) {
   }
 
   return parsed;
-}
-
-export function formatDailyCloseCompletedAt(completedAt?: number | null) {
-  if (!completedAt) return "Completion time unavailable";
-
-  return new Date(completedAt).toLocaleString([], {
-    day: "numeric",
-    hour: "numeric",
-    minute: "2-digit",
-    month: "short",
-    year: "numeric",
-  });
 }
 
 function getLocalOperatingDate(date = new Date()) {
