+++ packages/athena-webapp/src/components/orders/OrdersView.tsx
@@ -15,12 +15,19 @@ import { FadeIn } from "../common/FadeIn";
 import OrderMetricsPanel from "./OrderMetricsPanel";
 import { getAmountPaidForOrder } from "./utils";
 import { ProtectedRoute } from "../ProtectedRoute";
+import {
+  PageLevelHeader,
+  PageWorkspace,
+  PageWorkspaceMain,
+} from "../common/PageLevelHeader";
 
 type TimeRange = "day" | "week" | "month" | "all";
 
-export default function OrdersView({ status }: { status?: string }) {
+export default function OrdersView({ status = "all" }: { status?: string }) {
   const { activeStore } = useGetActiveStore();
-  const [selectedTimeRange, setSelectedTimeRange] = useState<TimeRange>("all");
+  const initialTimeRange: TimeRange = status === "all" ? "all" : "day";
+  const [selectedTimeRange, setSelectedTimeRange] =
+    useState<TimeRange>(initialTimeRange);
 
   const orders = useQuery(
     api.storeFront.onlineOrder.getAllOnlineOrders,
@@ -50,6 +57,10 @@ export default function OrdersView({ status }: { status?: string }) {
 
   const ordersFormatted = orders
     .filter((o: OnlineOrder) => {
+      if (timeFilter !== undefined && o._creationTime < timeFilter) {
+        return false;
+      }
+
       // Apply status filter based on the page
       if (status) {
         if (status == "completed") {
@@ -63,7 +74,7 @@ export default function OrdersView({ status }: { status?: string }) {
 
       return true;
     })
-    .map((order: any) => {
+    .map((order: OnlineOrder) => {
       // Calculate net amount (amount paid after discounts and including fees)
       const netAmount = getAmountPaidForOrder(order);
 
@@ -74,19 +85,10 @@ export default function OrdersView({ status }: { status?: string }) {
       };
     });
 
-  const hasOrders = ordersFormatted.length > 0;
-
-  const Navigation = () => {
-    return (
-      <div className="container mx-auto flex gap-2 h-[40px]">
-        <div className="flex items-center">
-          {status && hasOrders && (
-            <p className="text-xl font-medium">{`${capitalizeFirstLetter(slugToWords(status))} orders`}</p>
-          )}
-        </div>
-      </div>
-    );
-  };
+  const title =
+    status && status !== "all"
+      ? `${capitalizeFirstLetter(slugToWords(status))} orders`
+      : "Online orders";
 
   const handleTimeRangeChange = (timeRange: TimeRange) => {
     setSelectedTimeRange(timeRange);
@@ -94,27 +96,29 @@ export default function OrdersView({ status }: { status?: string }) {
 
   return (
     <ProtectedRoute requires="full_admin">
-      <View
-        hideBorder
-        hideHeaderBottomBorder
-        header={hasOrders && <Navigation />}
-      >
-        <OrderMetricsPanel
-          storeId={activeStore._id}
-          currency={activeStore.currency}
-          onTimeRangeChange={handleTimeRangeChange}
-        />
-        <div className="container mx-auto">
-          <FadeIn>
-            <OrdersTableToolbarProvider>
-              <Orders
-                store={activeStore}
-                status={status || "open"}
-                orders={ordersFormatted}
+      <View hideBorder hideHeaderBottomBorder scrollMode="page">
+        <FadeIn className="container mx-auto py-layout-xl">
+          <PageWorkspace>
+            <PageLevelHeader
+              eyebrow="Commerce Ops"
+              title={title}
+              description="Review online order volume, revenue, and fulfillment work without leaving the store operations flow."
+            />
+
+            <PageWorkspaceMain>
+              <OrderMetricsPanel
+                initialTimeRange={initialTimeRange}
+                storeId={activeStore._id}
+                currency={activeStore.currency}
+                onTimeRangeChange={handleTimeRangeChange}
               />
-            </OrdersTableToolbarProvider>
-          </FadeIn>
-        </div>
+
+              <OrdersTableToolbarProvider>
+                <Orders status={status} orders={ordersFormatted} />
+              </OrdersTableToolbarProvider>
+            </PageWorkspaceMain>
+          </PageWorkspace>
+        </FadeIn>
       </View>
     </ProtectedRoute>
   );
